refactor: migrate tests to external test packages within internal packages
