+++ internal/core/option.go
@@ -1,9 +1,7 @@
 package core
 
 import (
-	"fmt"
 	"net/url"
-	"strings"
 )
 
 const (
@@ -142,41 +140,4 @@ func HasRequiredOption(options []RequestOption, requiredTypes []APIParamOptionTy
 		}
 	}
 	return false
-}
-
-// InvalidOptionKeyError represents an error for an invalid option value.
-type InvalidOptionKeyError struct {
-	Invalid   string
-	ValidList []string
-}
-
-func NewInvalidOptionKeyError(invalid string, validList []APIParamOptionType) *InvalidOptionKeyError {
-	validKeys := []string{}
-	for _, v := range validList {
-		validKeys = append(validKeys, v.Value())
-	}
-
-	return &InvalidOptionKeyError{
-		Invalid:   invalid,
-		ValidList: validKeys,
-	}
-}
-
-func (e *InvalidOptionKeyError) Error() string {
-	return fmt.Sprintf("invalid option key:%s, allowed option keys:%s", e.Invalid, strings.Join(e.ValidList, ","))
-}
-
-// ValidationError represents an argument validation error.
-type ValidationError struct {
-	message string
-}
-
-func NewValidationError(msg string) *ValidationError {
-	return &ValidationError{
-		message: msg,
-	}
-}
-
-func (e *ValidationError) Error() string {
-	return e.message
 }
