+++ internal/testutil/mock/mock.go
@@ -28,6 +28,37 @@ func (m *MockDoer) Do(req *http.Request) (*http.Response, error) {
 }
 
 // ──────────────────────────────────────────────────────────────
+//  Wrapper mock
+// ──────────────────────────────────────────────────────────────
+
+type MockWrapper struct {
+	CreateErr error
+	CopyErr   error
+	CloseErr  error
+}
+
+func (w MockWrapper) NewMultipartWriter(_ io.Writer) core.MultipartWriter {
+	return &MockMultipartWriter{wrapper: w}
+}
+
+func (w MockWrapper) Copy(_ io.Writer, _ io.Reader) error {
+	return w.CopyErr
+}
+
+type MockMultipartWriter struct {
+	wrapper MockWrapper
+}
+
+func (mw *MockMultipartWriter) CreateFormFile(fieldname, filename string) (io.Writer, error) {
+	if mw.wrapper.CreateErr != nil {
+		return nil, mw.wrapper.CreateErr
+	}
+	return io.Discard, nil
+}
+func (mw *MockMultipartWriter) FormDataContentType() string { return "mock/type" }
+func (mw *MockMultipartWriter) Close() error                { return mw.wrapper.CloseErr }
+
+// ──────────────────────────────────────────────────────────────
 //  RequestOption mock
 // ──────────────────────────────────────────────────────────────
